+++ DelvUI/Config/ConfigurationManager.cs
@@ -309,6 +309,8 @@ namespace DelvUI.Config
         public ConfigPageNode GetConfigPageNode<T>() where T : PluginConfigObject => ConfigBaseNode.GetConfigPageNode<T>()!;
 
         public void SetConfigObject(PluginConfigObject configObject) => ConfigBaseNode.SetConfigObject(configObject);
+
+        public List<T> GetObjects<T>() => ConfigBaseNode.GetObjects<T>();
         #endregion
 
         #region load / save / profiles
@@ -347,7 +349,8 @@ namespace DelvUI.Config
 
         public void LoadConfigurations()
         {
-            ConfigBaseNode.Load(ConfigDirectory, CurrentVersion, PreviousVersion);
+            PerformV2Migration();
+            ConfigBaseNode.Load(ConfigDirectory);
         }
 
         public void SaveConfigurations(bool forced = false)
@@ -362,6 +365,49 @@ namespace DelvUI.Config
             ProfilesManager.Instance?.SaveCurrentProfile();
 
             ConfigBaseNode.NeedsSave = false;
+        }
+
+        public void PerformV2Migration()
+        {
+            // create necessary folders
+            string[] newFolders = new string[] { "Other Elements", "Customization" };
+            foreach (string folder in newFolders)
+            {
+                string path = Path.Combine(ConfigDirectory, folder);
+                if (!Directory.Exists(path))
+                {
+                    Directory.CreateDirectory(path);
+                }
+            }
+
+            // move files
+            Dictionary<string, string> files = new Dictionary<string, string>()
+            {
+                ["Misc\\Experience Bar.json"] = "Other Elements\\Experience Bar.json",
+                ["Misc\\GCD Indicator.json"]  = "Other Elements\\GCD Indicator.json",
+                ["Misc\\Limit Break.json"]    = "Other Elements\\Limit Break.json",
+                ["Misc\\MP Ticker.json"]      = "Other Elements\\MP Ticker.json",
+                ["Misc\\Pull Timer.json"]     = "Other Elements\\Pull Timer.json",
+                ["Misc\\Fonts.json"]          = "Customization\\Fonts.json"
+            };
+
+            foreach (string key in files.Keys)
+            {
+                string v1Path = Path.Combine(ConfigDirectory, key);
+                string v2Path = Path.Combine(ConfigDirectory, files[key]);
+
+                try
+                {
+                    if (File.Exists(v1Path))
+                    {
+                        File.Move(v1Path, v2Path);
+                    }
+                }
+                catch (Exception e)
+                {
+                    PluginLog.Error("Error migrating file \"" + v1Path + "\" to v2 config structure: " + e.Message);
+                }
+            }
         }
 
         public void UpdateCurrentProfile()
@@ -404,6 +450,8 @@ namespace DelvUI.Config
             ConfigBaseNode = node;
             ConfigBaseNode.ConfigObjectResetEvent += OnConfigObjectReset;
 
+            PerformV2Migration();
+
             ResetEvent?.Invoke(this);
 
             return true;
@@ -431,32 +479,6 @@ namespace DelvUI.Config
                 {
                     oldConfigObjects.Add(config.GetType(), config);
                 }
-            }
-
-            try
-            {
-                // handle imports for breaking changes in the config
-                if (UnmergeableConfigTypesPerVersion.TryGetValue(CurrentVersion, out List<Type>? types) && types != null)
-                {
-                    foreach (Type type in types)
-                    {
-                        var genericMethod = node.GetType().GetMethod("GetConfigObject");
-                        var method = genericMethod?.MakeGenericMethod(type);
-                        PluginConfigObject? config = (PluginConfigObject?)method?.Invoke(node, null);
-
-                        if (config != null)
-                        {
-                            config.ImportFromOldVersion(oldConfigObjects, CurrentVersion, PreviousVersion);
-                            node.SetConfigObject(config); // needed to refresh nodes
-                        }
-                    }
-                }
-
-                node.Save(ConfigDirectory);
-            }
-            catch
-            {
-                return false;
             }
 
             if (ProfilesManager.Instance != null)
@@ -488,21 +510,25 @@ namespace DelvUI.Config
 
         private static Type[] ConfigObjectTypes = new Type[]
         {
+            // Unit Frames
             typeof(PlayerUnitFrameConfig),
             typeof(TargetUnitFrameConfig),
             typeof(TargetOfTargetUnitFrameConfig),
             typeof(FocusTargetUnitFrameConfig),
 
+            // Mana Bars
             typeof(PlayerPrimaryResourceConfig),
             typeof(TargetPrimaryResourceConfig),
             typeof(TargetOfTargetPrimaryResourceConfig),
             typeof(FocusTargetPrimaryResourceConfig),
-
+            
+            // Castbars
             typeof(PlayerCastbarConfig),
             typeof(TargetCastbarConfig),
             typeof(TargetOfTargetCastbarConfig),
             typeof(FocusTargetCastbarConfig),
 
+            // Buffs and Debuffs
             typeof(PlayerBuffsListConfig),
             typeof(PlayerDebuffsListConfig),
             typeof(TargetBuffsListConfig),
@@ -511,6 +537,7 @@ namespace DelvUI.Config
             typeof(FocusTargetDebuffsListConfig),
             typeof(CustomEffectsListConfig),
 
+            // Nameplates
             typeof(NameplatesGeneralConfig),
             typeof(PlayerNameplateConfig),
             typeof(EnemyNameplateConfig),
@@ -523,6 +550,7 @@ namespace DelvUI.Config
             typeof(MinionNPCNameplateConfig),
             typeof(ObjectsNameplateConfig),
 
+            // Party Frames
             typeof(PartyFramesConfig),
             typeof(PartyFramesHealthBarsConfig),
             typeof(PartyFramesManaBarConfig),
@@ -532,10 +560,12 @@ namespace DelvUI.Config
             typeof(PartyFramesDebuffsConfig),
             typeof(PartyFramesTrackersConfig),
 
+            // Party Cooldowns
             typeof(PartyCooldownsConfig),
             typeof(PartyCooldownsBarConfig),
             typeof(PartyCooldownsDataConfig),
 
+            // Enemy List
             typeof(EnemyListConfig),
             typeof(EnemyListHealthBarConfig),
             typeof(EnemyListEnmityIconConfig),
@@ -544,6 +574,7 @@ namespace DelvUI.Config
             typeof(EnemyListBuffsConfig),
             typeof(EnemyListDebuffsConfig),
 
+            // Job Specific Bars
             typeof(PaladinConfig),
             typeof(WarriorConfig),
             typeof(DarkKnightConfig),
@@ -569,6 +600,14 @@ namespace DelvUI.Config
             typeof(RedMageConfig),
             typeof(BlueMageConfig),
 
+            // Other Elements
+            typeof(ExperienceBarConfig),
+            typeof(GCDIndicatorConfig),
+            typeof(PullTimerConfig),
+            typeof(LimitBreakConfig),
+            typeof(MPTickerConfig),
+
+            // Colors
             typeof(TanksColorConfig),
             typeof(HealersColorConfig),
             typeof(MeleeColorConfig),
@@ -577,33 +616,22 @@ namespace DelvUI.Config
             typeof(RolesColorConfig),
             typeof(MiscColorConfig),
 
+            // Customization
+            typeof(FontsConfig),
+            typeof(BarTexturesConfig),
+
+            // Visibility
             typeof(GlobalVisibilityConfig),
             typeof(HotbarsVisibilityConfig),
 
-            typeof(FontsConfig),
+            // Misc
             typeof(HUDOptionsConfig),
             typeof(WindowClippingConfig),
             typeof(TooltipsConfig),
-            typeof(ExperienceBarConfig),
-            typeof(GCDIndicatorConfig),
-            typeof(PullTimerConfig),
-            typeof(LimitBreakConfig),
-            typeof(MPTickerConfig),
             typeof(GridConfig),
 
+            // Import
             typeof(ImportConfig)
-        };
-
-        private static Dictionary<string, List<Type>> UnmergeableConfigTypesPerVersion = new Dictionary<string, List<Type>>()
-        {
-            ["0.4.0.0"] = new List<Type>() {
-                typeof(PartyFramesIconsConfig),
-                typeof(PartyFramesTrackersConfig)
-            },
-            ["0.6.2.0"] = new List<Type>() {
-                typeof(PartyFramesHealthBarsConfig)
-            },
-
         };
         #endregion
     }